# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CTEM Platform API — Continuous Threat Exposure Management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /findings:
    get:
      operationId: listFindings
      tags: [findings]
      summary: List vulnerability findings
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
        - name: assetId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of findings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Finding"
  /findings/{findingId}:
    get:
      operationId: getFinding
      tags: [findings]
      summary: Get finding details
      parameters:
        - name: findingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Finding details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
    patch:
      operationId: updateFinding
      tags: [findings]
      summary: Update finding status or details
      parameters:
        - name: findingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindingUpdate"
      responses:
        "200":
          description: Updated finding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
  /findings/{findingId}/comments:
    get:
      operationId: listFindingComments
      tags: [findings]
      summary: Get comments for a finding
      parameters:
        - name: findingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FindingComment"
    post:
      operationId: createFindingComment
      tags: [findings]
      summary: Add a comment to a finding
      parameters:
        - name: findingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindingCommentInput"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingComment"

  # ──────────────────────────────────────────────────────────
  # COMPLIANCE
  # ──────────────────────────────────────────────────────────
...
  /compliance/controls:
    get:
      operationId: listComplianceControls
      tags: [compliance]
      summary: List compliance controls
      parameters:
        - name: frameworkId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Compliance controls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplianceControl"
  /compliance/controls/{controlId}:
    get:
      operationId: getComplianceControl
      tags: [compliance]
      summary: Get control details
      parameters:
        - name: controlId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Control details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceControl"
    patch:
      operationId: updateComplianceControl
      tags: [compliance]
      summary: Update compliance control status
      parameters:
        - name: controlId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceControlUpdate"
      responses:
        "200":
          description: Updated control
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceControl"
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List generated reports
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Generate a new report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report creation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{reportId}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get report details
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────────────────────────────────────────────────
  # ALERTS
  # ──────────────────────────────────────────────────────────
...
  /risk/scores/{assetId}:
    get:
      operationId: getAssetRiskScore
      tags: [risk]
      summary: Get risk score for a specific asset
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskScore"

  # ──────────────────────────────────────────────────────────
  # AUDIT LOGS
  # ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Finding:
      type: object
      required: [id, tenantId, assetId, title, severity, status, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        assetId:
          type: integer
        assetName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        severity:
          type: string
        status:
          type: string
        cve:
          type: ["string", "null"]
        cvss:
          type: ["number", "null"]
        epss:
          type: ["number", "null"]
        cwe:
          type: ["string", "null"]
        isKev:
          type: boolean
        remediation:
          type: ["string", "null"]
        evidence:
          type: ["string", "null"]
        riskScore:
          type: ["number", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    FindingUpdate:
      type: object
      properties:
        status:
          type: string
        remediation:
          type: string
        evidence:
          type: string
    FindingComment:
      type: object
      required: [id, findingId, userId, content, createdAt]
      properties:
        id:
          type: integer
        findingId:
          type: integer
        userId:
          type: integer
        authorName:
          type: string
        content:
          type: string
        createdAt:
          type: string
    FindingCommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ── Compliance ────────────────────────────────────────
...
    ComplianceControl:
      type: object
      required: [id, frameworkId, frameworkName, controlId, title, status, createdAt]
      properties:
        id:
          type: integer
        frameworkId:
          type: integer
        frameworkName:
          type: string
        controlId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        evidence:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
    ComplianceControlUpdate:
      type: object
      properties:
        status:
          type: string
        evidence:
          type: string
        assignedTo:
          type: string
        dueDate:
          type: string
...
    Report:
      type: object
      required: [id, tenantId, title, type, format, status, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        title:
          type: string
        type:
          type: string
        format:
          type: string
        status:
          type: string
        downloadUrl:
          type: ["string", "null"]
        generatedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ReportInput:
      type: object
      required: [title, type, format]
      properties:
        title:
          type: string
        type:
          type: string
        format:
          type: string
        assetIds:
          type: array
          items:
            type: integer
        dateFrom:
          type: string
        dateTo:
          type: string

    # ── Alert ─────────────────────────────────────────────
...
    RiskScore:
      type: object
      required: [id, assetId, assetName, score, level, cvssComponent, epssComponent, kevBonus, criticalityBonus, exposureBonus, updatedAt]
      properties:
        id:
          type: integer
        assetId:
          type: integer
        assetName:
          type: string
        score:
          type: number
        level:
          type: string
        cvssComponent:
          type: number
        epssComponent:
          type: number
        kevBonus:
          type: number
        criticalityBonus:
          type: number
        exposureBonus:
          type: number
        updatedAt:
          type: string

    # ── Audit Log ─────────────────────────────────────────
...